Corregido DataLoader, agregados Repositorios y entidades faltantes

+++ src/main/java/com/example/clientes_venta/Compras/ComprasRepo.java
@@ -2,7 +2,7 @@ package com.example.clientes_venta.Compras;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ComprasRepo extends JpaRepository<Compras, Long>{
+public interface ComprasRepo extends JpaRepository<Compras, Integer>{
     public Compras findById(long id);
     
 }
