+++ src/main/java/com/example/clientes_venta/Producto/ProductoRepo.java
@@ -2,7 +2,7 @@ package com.example.clientes_venta.Producto;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
-public interface ProductoRepo extends JpaRepository<Producto, Long>{
+public interface ProductoRepo extends JpaRepository<Producto, Integer>{
 
     Optional<Producto> findByName(String name);
