+++ src/main/java/com/example/clientes_venta/Usuario/UsuarioRepo.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 
 
-public interface UsuarioRepo extends JpaRepository<Usuario, Long> {
+public interface UsuarioRepo extends JpaRepository<Usuario, Integer> {
 
     void deleteByName(String name);
