+++ src/main/java/com/example/clientes_venta/Ventas/VentasRepo.java
@@ -2,7 +2,6 @@ package com.example.clientes_venta.Ventas;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VentasRepo extends JpaRepository<Ventas, Long>{
-    public Ventas findById(long id);
-    
+public interface VentasRepo extends JpaRepository<Ventas, Integer>{ // <- CAMBIO AQUÍ
+    public Ventas findById(int id); // <- CAMBIO AQUÍ
 }
